Add function interfaceplot_advanced
A new interface plot function is added to the package. Like 'interfaceplots' this function uses the suggested package 'qgraph' to create the interface plots. Unlike 'interfaceplots' no interactive option is possible. However the new function allows for more flexibility and fine-tunning in the creation of the graphics.
A more detailed description of the function with examples will follow in future commits.
The new plot function is incorporated into the modules_interfaceplot function with the new input parameter 'use_advanced_interfacePlot_function'.